add_kafka
