--- feign-common-module/src/main/java/common/model/Book.java
@@ -1,16 +1,0 @@
-package common.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class Book {
-
-    String name;
-
-    String author;
-
-    Integer size;
-
-}
