--- feign-example-second-app/src/main/java/second/app/controller/BookController.java
@@ -1,28 +1,0 @@
-package second.app.controller;
-
-import common.model.Book;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/book")
-@Slf4j
-public class BookController {
-
-    @GetMapping("/getBook")
-    public Book getBook(){
-        log.info("getBook run");
-        return new Book("Random book","Random author",23);
-    }
-
-    @PostMapping("/createBook")
-    public Book createBook(@RequestBody Book book){
-        log.info("createBook run");
-        return book;
-    }
-
-}
