--- feign-example-second-app/src/main/resources/application.yml
@@ -1,10 +1,0 @@
-server:
-  port: 8485
-keycloak:
-  auth-server-url: ${KEYCLOAK_URL:http://localhost:8080/auth}
-  realm: ${KEYCLOAK_REALM:my_realm}
-  resource: ${KEYCLOAK_RESOURCE:my_client}
-  public-client: true
-spring:
-  application:
-    name: ${APPLICATION_NAME:Feign-Example}
