+++ feign-kafka-rmi-common-module/dto/src/main/java/common/dto/Book.java
@@ -1,0 +1,18 @@
+package common.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Book {
+
+    String name;
+
+    String author;
+
+    Integer size;
+
+}
