+++ feign-kafka-rmi-common-module/dto/src/main/java/common/dto/FoodOrder.java
@@ -1,0 +1,13 @@
+package common.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class FoodOrder {
+    String item;
+    Double amount;
+}
