+++ feign-kafka-rmi-first-app/src/main/java/first/app/kafka/Producer.java
@@ -1,0 +1,31 @@
+package first.app.kafka;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import common.dto.FoodOrder;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class Producer {
+
+    @Value("${topic.name}")
+    private String orderTopic;
+
+    private final ObjectMapper objectMapper;
+    private final KafkaTemplate<String, String> kafkaTemplate;
+
+    public String sendMessage(FoodOrder foodOrder) throws JsonProcessingException {
+        String orderAsMessage = objectMapper.writeValueAsString(foodOrder);
+        kafkaTemplate.send(orderTopic, orderAsMessage);
+
+        log.info("food order produced {}", orderAsMessage);
+
+        return "message sent";
+    }
+}
