+++ feign-kafka-rmi-first-app/src/main/java/first/app/service/FoodOrderService.java
@@ -1,0 +1,20 @@
+package first.app.service;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import common.dto.FoodOrder;
+import first.app.kafka.Producer;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class FoodOrderService {
+
+    private final Producer producer;
+
+    public String createFoodOrder(FoodOrder foodOrder) throws JsonProcessingException {
+        return producer.sendMessage(foodOrder);
+    }
+}
