+++ feign-kafka-rmi-first-app/src/main/resources/application.yml
@@ -1,0 +1,4 @@
+server:
+  port: 8484
+topic:
+  name: t.food.order
