+++ feign-kafka-rmi-second-app/src/main/java/second/app/kafka/Consumer.java
@@ -1,0 +1,30 @@
+package second.app.kafka;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import common.dto.FoodOrder;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.stereotype.Component;
+
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class Consumer {
+
+    private static final String orderTopic = "${topic.name}";
+
+    private final ObjectMapper objectMapper;
+
+    @KafkaListener(topics = orderTopic)
+    public void consumeMessage(String message) throws JsonProcessingException {
+        log.info("message consumed {}", message);
+
+        FoodOrder foodOrder = objectMapper.readValue(message, FoodOrder.class);
+
+        log.info("New order is ready: " + foodOrder.getItem() + ", count: " + foodOrder.getAmount());
+    }
+
+}
